+++ projects/ngrx-clean-forms/src/public-api.ts
@@ -1,4 +1,5 @@
 export { initFormControl, initFormGroup, initFormArray } from './lib/init';
+export { resetFormControl, resetFormGroup, resetFormArray } from './lib/reset';
 export { reduceFormControl, reduceFormGroup, reduceFormArray } from './lib/reducer';
 export { getFormControlSummary, getFormGroupSummary, getFormArraySummary } from './lib/selectors';
 export * from './lib/types';